+++ internal/bootcamp/feedbackclient/feedbackclient.go
@@ -3,19 +3,19 @@ package feedbackclient
 import (
 	"context"
 	"google.golang.org/grpc"
+	"grabvn-golang-bootcamp/internal/bootcamp/configuration"
 	"grabvn-golang-bootcamp/internal/bootcamp/feedback"
 	"log"
 	"time"
 )
 
-const (
-	port = ":9000"
-)
-
 func StartClient() {
+	var config configuration.Conf
+	config.LoadConf()
+
 	log.Print("begin init rpc client....")
 
-	conn, err := grpc.Dial("localhost"+port, grpc.WithInsecure())
+	conn, err := grpc.Dial("localhost"+config.RPC.Port, grpc.WithInsecure())
 
 	if err != nil {
 		log.Fatalf("failed to connect server: %v", err)
@@ -25,7 +25,6 @@ func StartClient() {
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
-
 
 	res, err := client.GetById(ctx, &feedback.FeedbackRequest{})
 	if err != nil {
