+++ internal/bootcamp/feedbackserver/server.go
@@ -23,10 +23,10 @@ func (s *server) Add(ctx context.Context, in *feedback.CreateFeedbackRequest) (*
 
 	err := s.DB.Create(&res).Error
 	if err == nil {
-		return &feedback.FeedbackResponse{ID: res.ID, Status: feedback.ResponseStatus_SUCCESS}, nil
+		return &feedback.FeedbackResponse{ID: res.ID}, nil
 	} else {
 		log.Print(err)
-		return &feedback.FeedbackResponse{ID: res.ID, Status: feedback.ResponseStatus_FAIL}, nil
+		return &feedback.FeedbackResponse{ID: res.ID}, nil
 	}
 }
 
@@ -46,37 +46,29 @@ func (s *server) GetById(ctx context.Context, in *feedback.FeedbackRequest) (*fe
 		return &feedback.PassengerFeedback{ID: in.ID}, errors.New("not found")
 	}
 }
-func (s *server) GetByBookingCode(ctx context.Context, in *feedback.BookingRequest) (*feedback.ListFeedbackResponse, error) {
-	var res []Feedback
+func (s *server) GetByBookingCode(ctx context.Context, in *feedback.BookingRequest) (*feedback.PassengerFeedback, error) {
+	var res Feedback
 	err := s.DB.Where("BookingCode = ?", in.Code).Find(&res).GetErrors()
 	if len(err) == 0 {
-		var fb []*feedback.PassengerFeedback
-
-		for _, f := range res {
-			fb = append(fb, &feedback.PassengerFeedback{
-				ID:          f.ID,
-				PassengerID: f.PassengerID,
-				BookingCode: f.BookingCode,
-				Feedback:    f.Feedback,
-			})
+		fb := feedback.PassengerFeedback{
+			ID:          res.ID,
+			PassengerID: res.PassengerID,
+			BookingCode: res.BookingCode,
+			Feedback:    res.Feedback,
 		}
-
-		data := feedback.ListFeedbackResponse{
-			Status:   feedback.ResponseStatus_SUCCESS,
-			Feedback: fb,
-		}
-		return &data, nil
+		return &fb, nil
 
 	} else {
-		return &feedback.ListFeedbackResponse{Status: feedback.ResponseStatus_FAIL}, errors.New("not found")
+		return &feedback.PassengerFeedback{BookingCode: in.Code}, errors.New("not found")
 	}
+
 }
 func (s *server) Delete(ctx context.Context, in *feedback.FeedbackRequest) (*feedback.FeedbackResponse, error) {
 	err := s.DB.Where("ID = ?", in.ID).Delete(Feedback{}).GetErrors()
 	if len(err) == 0 {
-		return &feedback.FeedbackResponse{ID: in.ID, Status: feedback.ResponseStatus_SUCCESS}, nil
+		return &feedback.FeedbackResponse{ID: in.ID}, nil
 	} else {
-		return &feedback.FeedbackResponse{ID: in.ID, Status: feedback.ResponseStatus_FAIL}, errors.New("internal found")
+		return &feedback.FeedbackResponse{ID: in.ID}, errors.New("internal found")
 	}
 
 }
